+++ packages/next-common/components/fellowship/salary/common.js
@@ -1,4 +1,5 @@
 import ListLayout from "next-common/components/layout/ListLayout";
+import FellowshipSalarySummary from "./summary";
 
 export default function FellowshipSalaryCommon({ children, ...props }) {
   const title = "Fellowship Salary";
@@ -11,10 +12,16 @@ export default function FellowshipSalaryCommon({ children, ...props }) {
       seoInfo={seoInfo}
       title={title}
       description={seoInfo.desc}
+      summary={<FellowshipSalarySummary />}
       tabs={[
         {
           label: "Cycles",
           url: "/fellowship/salary",
+          exactMatch: true,
+        },
+        {
+          label: "Claimants",
+          url: "/fellowship/salary/claimants",
           exactMatch: true,
         },
         {