+++ packages/next-common/components/popup/wrapper/Popup.js
@@ -64,7 +64,7 @@ export default function Popup({
                   <div className="flex items-center gap-[12px]">
                     {extra}
                     <SystemClose
-                      className="w-[20px] h-[20px]"
+                      className="w-[20px] h-[20px] text-textTertiary"
                       role="button"
                       onClick={onClose}
                     />