+++ packages/next/pages/fellowship/salary/feeds.js
@@ -1,9 +1,8 @@
-import { withCommonProps } from "next-common/lib";
-import { fetchOpenGovTracksProps } from "next-common/services/serverSide";
 import nextApi from "next-common/services/nextApi";
 import { defaultPageSize } from "next-common/utils/constants";
 import FellowshipSalaryCommon from "next-common/components/fellowship/salary/common";
 import FellowshipSalaryFeedsContainer from "next-common/components/fellowship/salary/feeds/container";
+import { withFellowshipSalaryCommonProps } from "next-common/services/serverSide/fellowship/common";
 
 export default function FellowshipSalaryFeedsPage({ fellowshipSalaryFeeds }) {
   return (
@@ -13,21 +12,22 @@ export default function FellowshipSalaryFeedsPage({ fellowshipSalaryFeeds }) {
   );
 }
 
-export const getServerSideProps = withCommonProps(async (context) => {
-  const { page = 0 } = context.query;
+export const getServerSideProps = withFellowshipSalaryCommonProps(
+  async (context) => {
+    const { page = 0 } = context.query;
 
-  const [tracksProps, { result: fellowshipSalaryFeeds }] = await Promise.all([
-    fetchOpenGovTracksProps(),
-    nextApi.fetch("fellowship/salary/feeds", {
-      page,
-      page_size: defaultPageSize,
-    }),
-  ]);
+    const { result: fellowshipSalaryFeeds } = await nextApi.fetch(
+      "fellowship/salary/feeds",
+      {
+        page,
+        page_size: defaultPageSize,
+      },
+    );
 
-  return {
-    props: {
-      ...tracksProps,
-      fellowshipSalaryFeeds,
-    },
-  };
-});
+    return {
+      props: {
+        fellowshipSalaryFeeds: fellowshipSalaryFeeds || {},
+      },
+    };
+  },
+);